# openapi.yaml
openapi: 3.1.0
info:
  title: Eka Legal Consultancy API
  description: |
    Backend API for Eka Legal Consultancy practice management system.
    Supports client portal, booking system, case management, messaging, and document handling.
  version: 1.0.0
  contact:
    name: Eka Legal Support
    email: support@eka-legal.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.eka-legal.com/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    LoginCredentials:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
      required:
        - email
        - password
    RegisterData:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        phone:
          type: string
      required:
        - name
        - email
        - password
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              description: JWT access token
        message:
          type: string

    # ============================================
    # Lawyer Profile & Public Content
    # ============================================
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCredentials'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new client account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterData'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate user session
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get the currently authenticated user's profile
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send password reset email to user
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
                - email
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset password using token from email
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                password:
                  type: string
                  format: password
                  minLength: 8
              required:
                - token
                - password
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid or expired token

  # ============================================
  # Public Endpoints
  # ============================================
...
  /public/contact:
    post:
      tags:
        - Public
      summary: Submit contact form
      description: Submit a contact inquiry
      operationId: submitContactForm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactFormData'
      responses:
        '200':
          description: Contact form submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================
  # Booking Endpoints
  # ============================================
...
  /booking/available-slots:
    get:
      tags:
        - Booking
      summary: Get available time slots
      description: Retrieve available booking slots for a specific date
      operationId: getAvailableSlots
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Date to check availability (YYYY-MM-DD)
      responses:
        '200':
          description: Available slots retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TimeSlot'
  /booking/bookings:
    get:
      tags:
        - Booking
      summary: Get my bookings
      description: Retrieve all bookings for the authenticated user
      operationId: getMyBookings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Bookings retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Booking'
        '401':
          description: Not authenticated

    post:
      tags:
        - Booking
      summary: Create booking
      description: Create a new consultation booking
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Booking'
        '400':
          description: Invalid booking request
  /booking/bookings/{bookingId}:
    delete:
      tags:
        - Booking
      summary: Cancel booking
      description: Cancel an existing booking
      operationId: cancelBooking
      security:
        - BearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Booking not found

  # ============================================
  # Case Endpoints
  # ============================================
  /cases:
    get:
      tags:
        - Cases
      summary: Get my cases
      description: Retrieve all cases for the authenticated user
      operationId: getMyCases
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, active, closed]
          description: Filter cases by status
      responses:
        '200':
          description: Cases retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Case'
        '401':
          description: Not authenticated
  /cases/{caseId}:
    get:
      tags:
        - Cases
      summary: Get case by ID
      description: Retrieve a specific case by its ID
      operationId: getCaseById
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Case retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Case'
        '404':
          description: Case not found

  # ============================================
  # Document Endpoints
  # ============================================
...
  /messages/conversations:
    get:
      tags:
        - Messages
      summary: Get conversations
      description: Retrieve all conversations for the authenticated user
      operationId: getConversations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Conversations retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Conversation'
  /messages/conversations/{conversationId}/messages:
    get:
      tags:
        - Messages
      summary: Get messages
      description: Retrieve messages for a specific conversation
      operationId: getMessages
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Message'

    post:
      tags:
        - Messages
      summary: Send message
      description: Send a message in a conversation
      operationId: sendMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
          multipart/form-data:
            schema:
              type: object
              properties:
                content:
                  type: string
                attachments:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - content
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Message'
  /messages/read:
    post:
      tags:
        - Messages
      summary: Mark messages as read
      description: Mark multiple messages as read
      operationId: markMessagesAsRead
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messageIds:
                  type: array
                  items:
                    type: string
              required:
                - messageIds
      responses:
        '200':
          description: Messages marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================
  # Notification Endpoints
  # ============================================
  /notifications:
    get:
      tags:
        - Notifications
      summary: Get notifications
      description: Retrieve all notifications for the authenticated user
      operationId: getNotifications
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Notifications retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
  /notifications/{notificationId}/read:
    post:
      tags:
        - Notifications
      summary: Mark notification as read
      description: Mark a specific notification as read
      operationId: markNotificationAsRead
      security:
        - BearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /notifications/read-all:
    post:
      tags:
        - Notifications
      summary: Mark all as read
      description: Mark all notifications as read
      operationId: markAllNotificationsAsRead
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================
  # Dashboard Endpoints
  # ============================================
  /dashboard/client/stats:
    get:
      tags:
        - Dashboard
      summary: Get client dashboard stats
      description: Retrieve dashboard statistics for clients
      operationId: getClientStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Client stats retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DashboardStats'
...